Prepare to create database
