+++ POEItemFilter/POEItemFilter.csproj
@@ -202,7 +202,9 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
+    <Compile Include="Models\ApplicationDbContext.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\ItemsDB\Enum\Attribute.cs" />
     <Compile Include="Models\ItemsDB\Enum\BaseType.cs" />
