+++ POEItemFilter/Web.config
@@ -9,7 +9,9 @@
     <section name="entityFramework" type="System.Data.Entity.Internal.ConfigFile.EntityFrameworkSection, EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" requirePermission="false" />
   </configSections>
   <connectionStrings>
-    <add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-POEItemFilter-20171009122337.mdf;Initial Catalog=aspnet-POEItemFilter-20171009122337;Integrated Security=True" providerName="System.Data.SqlClient" />
+    <add name="POEItemFilterContext" 
+         connectionString="Data Source=.\SQLEXPRESS;AttachDbFilename=|DataDirectory|\POEItemFilter.mdf;Initial Catalog=POEItemFilter;Integrated Security=True;MultipleActiveResultSets=True;App=EntityFramework" 
+         providerName="System.Data.SqlClient" />
   </connectionStrings>
   <appSettings>
     <add key="webpages:Version" value="3.0.0.0" />
